+++ CHANGELOG.md
@@ -19,6 +19,7 @@ QML Material CHANGELOG
  * [Dialog] Add support for full-width content
  * [Icons] Update the FontAwesome parsing script to pull from the website
  * [ListItem] Correctly calculate the width of a simple menu item
+ * [ActionBar] Hide non-visible actions in the action bar
 
 
 ### Version 0.0.6 (April 16, 2015)